made button reusable
